Add --json flag
Allows input to be pre-structured as JSON. Applications that can output JSON
natively can forgoe the complexity and overhead of grokking with regexps.

This is an extension to `io.messages` with some basic validation. Each event
has to be on a single line, otherwise it's not possible to determine where a
stream ends.