Don't overwrite @timestamp
Applications that create their own JSON structured events will already
include a timestamp field. This reflects the time that the event happened,
which is more important than the time logship consumes it.

It does not however perform any normalisation. Which may be an issue,
because not many things than generate ISO8601 timestamps with u-second
precision.