[#46836025] Remove --no-stamp logship integ test
Remove the `--no-stamp` option from the logship integration test when
passing in JSON. This makes little sense for logship, since we always want
to ship with a timestamp. We also need to ensure that a TS in JSON is
preserved.

Split timestamp generation (when JSON doesn't contain TS) out to a seaparate
test which is marked as skipped until we can figure a way to mock the call
to `tagalog._now()`.