Integration test for unicode JSON
I was reasonably confident that the JSON input would accept unicode because
it was reusing `io.lines()`. However this integration test proves it.

I did attempt to `readlines()` on the fixture and construct each line of
JSON as a concatenated string. But this was clearly quite crazy, when we
already have the utilities to turn a message into JSON and then strip it
back down again. The comparison of `bytes()` objects should be much more
reliable.