[#46898331] Add the source host to the message
This change passes the integration tests. However, once we do the next stage, i.e. having Tagalog calculate the source host, then this will become problematic as we cannot mock this in the test. One possibility is adding a --no-source-host flag, as per --no-stamp; another is refactoring the source_host method to do less. @dcarley and I will discuss tomorrow.