[#46898331] Simplifying to cover cmd line only
The reason we had to do this now is because it is blocking https://www.pivotaltracker.com/story/show/45081601. However, in order to move onto that we only need to make Tagalog take the source-host command line argument - we do not need to have Tagalog calculate the source host. That is actually not trivial, so I am separating that into a different story so we can continue with the higher priority https://www.pivotaltracker.com/story/show/45081601.