Fix major bug: connections leaked on RedisError
If a RedisError was thrown by the underlying redis package, Tagalog would catch
the error and print a diagnostic message to its logger. Unfortunately, it would
leak the connection in the process, never returning it to the redis connection
pool.

This commit makes sure that if *any* exception is thrown while handling a
connection, that connection is purged from the connection pool.

If you've been experiencing large numbers of connections in CLOSE_WAIT being
held by the logship python processes, this change should hopefully fix that.